+++ docs/conf.py
@@ -61,6 +61,10 @@ nitpick_ignore = [
     ("py:class", "SimTime"),
     ("py:class", "immutables._map.Map"),
     ("py:class", "_asyncio.Task"),
+    ("py:class", "asyncio.streams.StreamReader"),
+    ("py:class", "asyncio.streams.StreamWriter"),
+    ("py:class", "apischema.conversions.conversions.Conversion"),
+    ("py:class", "apischema.conversions.conversions.LazyConversion"),
 ]
 
 # Both the class’ and the __init__ method’s docstring are concatenated and