+++ tickit/devices/femto/femto.py
@@ -1,23 +1,23 @@
-from random import uniform
-from typing import Awaitable, Callable, Dict
-
 from softioc import builder
 
-from tickit.adapters.epicsadapter import EpicsAdapter, InputRecord, OutputRecord
-from tickit.core.device import ConfigurableDevice, DeviceUpdate
-from tickit.core.typedefs import SimTime, State
+from tickit.adapters.epicsadapter import EpicsAdapter
+from tickit.core.device import Device, DeviceUpdate
+from tickit.core.typedefs import SimTime
 from tickit.utils.compat.typing_compat import TypedDict
 
 
-class Femto(ConfigurableDevice):
+class FemtoDevice(Device):
     """Electronic signal amplifier."""
 
-    Output = TypedDict("Output", {"current": float})
+    #: An empty typed mapping of device inputs
+    Inputs: TypedDict = TypedDict("Inputs", {"input": float})
+    #: A typed mapping containing the current output value
+    Outputs: TypedDict = TypedDict("Outputs", {"current": float})
 
     def __init__(
         self,
-        initial_gain: float = 2.5,
-        initial_current: float = 0.0,
+        initial_gain: float,
+        initial_current: float,
     ) -> None:
         """Initialise the Femto device class.
 
@@ -46,13 +46,16 @@ class Femto(ConfigurableDevice):
         """
         return self.gain
 
-    def set_current(self, current: float) -> None:
-        """Set the current of the output signal.
+    def set_current(self, input_current: float) -> None:
+        """Set the output current based on the input current and the gain.
+
+        The output current is calculated as:
+            output_current = input_current * gain
 
         Args:
-            current (float): The current to set the output signal to.
+            input_current (float): The current of the input signal.
         """
-        self._current = current * self.gain
+        self._output_current = input_current * self.gain
 
     def get_current(self) -> float:
         """Returns the output current of the signal.
@@ -60,10 +63,9 @@ class Femto(ConfigurableDevice):
         Returns:
             float: The output current of the signal.
         """
-        return self._current
+        return self._output_current
 
-    # Changed State for dict
-    def update(self, time: SimTime, inputs: dict) -> DeviceUpdate:
+    def update(self, time: SimTime, inputs: Inputs) -> DeviceUpdate[Outputs]:
         """Updates the state of the Femto device.
 
         Args:
@@ -73,79 +75,17 @@ class Femto(ConfigurableDevice):
         Returns:
             DeviceUpdate: A container for the Device's outputs and a callback time.
         """
-        current_value = inputs.get("input", None)
+        current_value = inputs["input"]
         if current_value is not None:
             self.set_current(current_value)
 
-        return DeviceUpdate(Femto.Output(current=self.gain), None)
-
-
-class CurrentDevice(ConfigurableDevice):
-    """The current configured device."""
-
-    Output = TypedDict("Output", {"output": float})
-
-    def __init__(self, callback_period: int = int(1e9)) -> None:
-        """Initialise the current device.
-
-        Args:
-            callback_period (Optional[int]): The duration in which the device should \
-                next be updated. Defaults to int(1e9).
-        """
-        self.callback_period = SimTime(callback_period)
-
-    def update(self, time: SimTime, inputs: State) -> DeviceUpdate:
-        """Updates the state of the current device.
-
-        Args:
-            time (SimTime): The time of the simulation in nanoseconds.
-            inputs (State): The state of the input values of the device.
-
-        Returns:
-            DeviceUpdate: A container for the Device's outputs and a callback time.
-        """
-        output = uniform(0.1, 200.1)
-        print(
-            "Output! (delta: {}, inputs: {}, output: {})".format(time, inputs, output)
-        )
-        return DeviceUpdate(
-            CurrentDevice.Output(output=output), SimTime(time + self.callback_period)
-        )
+        return DeviceUpdate(self.Outputs(current=self.get_current()), None)
 
 
 class FemtoAdapter(EpicsAdapter):
     """The adapter for the Femto device."""
 
-    current_record: InputRecord
-    input_record: InputRecord
-    output_record: OutputRecord
-
-    def __init__(
-        self,
-        device: Femto,
-        raise_interrupt: Callable[[], Awaitable[None]],
-        db_file: str = "record.db",
-        ioc_name: str = "FEMTO",
-    ):
-        """Initialise the Femto device adapter.
-
-        Args:
-            device (Device): The Femto device class.
-            raise_interrupt (Callable[[], Awaitable[None]]): A method used to request \
-                an immediate update of the device.
-            db_file (Optional[str]): The name of the database file. \
-                Defaults to "record.db".
-            ioc_name (Optional[str]): The name of the IOC. Defaults to "FEMTO".
-        """
-        super().__init__(db_file, ioc_name)
-        self._device = device
-        self.raise_interrupt = raise_interrupt
-
-        self.interrupt_records: Dict[InputRecord, Callable] = {}
-
-    async def run_forever(self) -> None:
-        """Builds the IOC."""
-        self.build_ioc()
+    device: FemtoDevice
 
     async def callback(self, value) -> None:
         """Device callback function.
@@ -153,18 +93,13 @@ class FemtoAdapter(EpicsAdapter):
         Args:
             value (float): The value to set the gain to.
         """
-        print("Callback", value)
-        self._device.set_gain(value)
+        self.device.set_gain(value)
         await self.raise_interrupt()
 
     def on_db_load(self) -> None:
         """Customises records that have been loaded in to suit the simulation."""
-        self.input_record = builder.aIn("GAIN_RBV")
-        self.output_record = builder.aOut(
-            "GAIN", initial_value=self._device.get_gain(), on_update=self.callback
+        builder.aOut(
+            "GAIN", initial_value=self.device.get_gain(), on_update=self.callback
         )
-
-        self.current_record = builder.aIn("CURRENT")
-
-        self.link_input_on_interrupt(self.input_record, self._device.get_gain)
-        self.link_input_on_interrupt(self.current_record, self._device.get_current)
+        self.link_input_on_interrupt(builder.aIn("GAIN_RBV"), self.device.get_gain)
+        self.link_input_on_interrupt(builder.aIn("CURRENT"), self.device.get_current)